2303-capstone-group-C/src/etl_service/requirements.txt
py4j==0.10.9.7
pyspark==3.4.0
python-decouple==3.8
redis==4.5.5
requests==2.31.0
